Querying is now a subcmd (geoid query)
