+++ geoid/cli/commands/query.py
@@ -3,17 +3,62 @@ from selenium.common.exceptions import *
 from geoid.bigquery import BigQuery
 from geoid.common import webclient
 from geoid.config import Config
-from geoid.logging import LOG_CONFIG
-import logging, logging.config, os
+from geoid.logging import log_std
+import logging, logging.config, time
 
 
 logger = logging.getLogger(__name__)
 
 
-def _start_logging():
-  if not os.path.exists(r'./logs/'):
-    os.mkdir(r'./logs/')
-  logging.config.dictConfig(LOG_CONFIG)
+def start_query(args):
+  if args.timestamp:
+    if "{timestamp}" not in args.output:
+      print(
+        'Specify timestamp location in the output filename using {timestamp} '
+        'to use the timestamp option'
+      )
+      return
+    else:      
+      timestamp = time.strftime('%Y%m%d_%H%M%S')
+      output_file = args.output.replace("{timestamp}", timestamp)
+  else:
+    output_file = args.output
+  
+  config = Config()
+  config.query.depth                 = args.depth
+  config.fileio.output_indent        = args.indent
+  config.webclient.webclient         = args.browser
+  config.query.initial_pause_seconds = args.init_pause
+  config.webclient.show              = args.show
+  config.fileio.use_timestamp_name   = args.timestamp
+  config.fileio.keep_autosave        = args.keep_autosave
+  config.postproc.filter             = args.filter
+  config.postproc.flatten            = args.flatten
+  config.postproc.convert_ascii      = args.convert_ascii
+  config.postproc.replace_newline    = args.replace_newline
+
+  if args.cities_file is None and args.cities is not None:
+    print(
+      f'Starting query, source: list.\n'
+      f'Query keyword: "{args.term} <cityname>"'
+    )
+    run_list(
+      term=args.term,
+      cities=args.cities,
+      output_file=output_file,
+      use_config=config
+    )
+  elif args.cities_file is not None and args.cities is None:
+    print(
+      f'Starting query, source: file.\n'
+      f'Query keyword: "{args.term} <cityname>"'
+    )
+    run_batch(
+      term=args.term,
+      source_file=args.cities_file,
+      output_file=output_file,
+      use_config=config
+    )
 
 
 def run_batch(
@@ -48,7 +93,7 @@ def run_batch(
       settings.
   """
 
-  _start_logging()
+  log_std()
   config = use_config if use_config else Config()
 
   querier = BigQuery(config)
@@ -91,7 +136,7 @@ def run_list(
       settings.
   """
 
-  _start_logging()
+  log_std
   config = use_config if use_config else Config()
 
   querier = BigQuery(config)
